requirements.txt
# Core data processing
pandas==2.3.0
numpy==2.3.1

# Data visualization
matplotlib==3.10.3
plotly==6.2.0
mplfinance==0.12.10b0

# Web scraping and HTTP requests
requests==2.32.4
beautifulsoup4==4.13.4
selenium==4.33.0
webdriver-manager==4.0.2
lxml==5.4.0

# Financial data APIs
yfinance==0.2.63
investpy==1.0.8

# AI/ML for commentary generation
google-generativeai==0.8.5

# Template rendering and markdown
Jinja2==3.1.6
Markdown==3.8.2

# Configuration and data handling
python-dotenv==1.1.1
pydantic==2.11.7
PyYAML==6.0.2

# Retry logic and robustness
tenacity==9.1.0

# Date and time handling
python-dateutil==2.9.0.post0
pytz==2025.2

# Progress bars and utilities
tqdm==4.67.1

# System monitoring
psutil==6.2.0

# Development and testing dependencies (optional)
# Uncomment these for development:
# pytest==8.0.0
# pytest-mock==3.14.0
# pytest-cov==5.0.0

# Linting and formatting (optional)
# Uncomment these for development:
# flake8==7.0.0
# black==24.0.0
